Add back nth_prime_sieve1()
