+++ src/print.cpp
@@ -295,13 +295,30 @@ void print_gourdon_vars(maxint_t x, int64_t y, int64_t z, int64_t k, int threads
   }
 }
 
-void print_nth_prime_sieve(uint64_t n,
-                           bool sieve_forward,
-                           maxint_t nth_prime_approx,
-                           uint64_t dist_approx,
-                           uint64_t thread_dist,
-                           int main_threads,
-                           int threads_per_segment)
+void print_nth_prime_sieve1(uint64_t n,
+                            bool sieve_forward,
+                            maxint_t nth_prime_approx,
+                            uint64_t dist_approx,
+                            uint64_t thread_dist,
+                            int threads)
+{
+  std::cout << "n = " << n << std::endl;
+  std::cout << "sieve_forward = " << (sieve_forward ? "true" : "false") << std::endl;
+  std::cout << "nth_prime_approx = " << nth_prime_approx << std::endl;
+  std::cout << "dist_approx = " << dist_approx << std::endl;
+  std::cout << "thread_dist = " << thread_dist << std::endl;
+  std::cout << "threads = " << threads << std::endl;
+}
+
+#if _OPENMP >= 201307
+
+void print_nth_prime_sieve2(uint64_t n,
+                            bool sieve_forward,
+                            maxint_t nth_prime_approx,
+                            uint64_t dist_approx,
+                            uint64_t thread_dist,
+                            int main_threads,
+                            int threads_per_segment)
 {
   std::cout << "n = " << n << std::endl;
   std::cout << "sieve_forward = " << (sieve_forward ? "true" : "false") << std::endl;
@@ -311,5 +328,7 @@ void print_nth_prime_sieve(uint64_t n,
   std::cout << "main_threads = " << main_threads << std::endl;
   std::cout << "threads_per_segment = " << threads_per_segment << std::endl;
 }
+
+#endif
 
 } // namespace
